tsa-auth-backend/requirements.txt
boto3>=1.26.0
botocore>=1.29.0
sendgrid==6.10.0
python-http-client==3.3.7 
PyJWT>=2.8.0 